[Testing: Workouts] Workouts Unit Tests (#13)
* Starting Workout Unit Tests

* Unit Testing Refactoring

* Refactoring of original unit tests

* Workouts Unit Tests

* Starting Exercise Unit Tests

* Unit Tests Rough Draft

* Refactoring Unit Tests

* Unit Testing Refactoring Rough Draft

* Updating lint/tests package.json methods

* Shared methods and constants for unit testing

* Workouts Unit Tests

+++ lib/home/workouts/workouts.ts
@@ -1,5 +1,5 @@
 "use server";
-import prisma from "@/client";
+import prisma from "@/lib/prisma/client";
 import { z } from "zod";
 import {
    sendSuccessMessage,
@@ -10,6 +10,7 @@ import {
 import { formatWorkout } from "@/lib/home/workouts/shared";
 import { Exercise } from "@/lib/home/workouts/exercises";
 import { uuidSchema } from "@/lib/global/zod";
+import { getAppliedWorkoutTagUpdates } from "@/lib/home/workouts/tags";
 
 export type Workout = {
   id: string;
@@ -27,38 +28,42 @@ const nextMediaRegex =
   /^\/workouts\/(bike|cardio|default|hike|legs|lift|machine|run|swim|weights)\.png$/;
 
 const workoutsSchema = z.object({
-   user_id: uuidSchema,
-   id: uuidSchema,
+   user_id: uuidSchema("user", "required"),
+   id: uuidSchema("workout", "required"),
    title: z
       .string()
       .trim()
-      .min(1, { message: "A title must be at least 1 character" })
-      .max(50, { message: "A title must be at most 50 characters" }),
+      .min(1, { message: "Title must be at least 1 character" })
+      .max(50, { message: "Title must be at most 50 characters" }),
    date: z
       .date({
-         required_error: "Date for workout is required",
-         invalid_type_error: "A valid date is required"
+         required_error: "Date is required",
+         invalid_type_error: "Date is required"
       })
       .max(new Date(new Date().getTime() + 24 * 60 * 60 * 1000), {
-         message: "A workout date must not be after today"
+         message: "Date must not be after today"
       }),
    description: z.string().optional().or(z.literal("")),
    image: z
       .string()
       .refine((value) => urlRegex.test(value) || nextMediaRegex.test(value), {
-         message: "Invalid URL"
-      })
-      .or(z.literal("")),
+         message: "Image URL must be valid"
+      }).or(z.literal("")),
    tags: z.array(z.string()).optional()
 });
 
-export async function fetchWorkouts(userId: string): Promise<Workout[]> {
+const newWorkoutSchema = workoutsSchema.extend({
+   id: uuidSchema("workout", "new")
+});
+
+export async function fetchWorkouts(user_id: string): Promise<Workout[]> {
    try {
-      const workoutsWithTags = await prisma.workouts.findMany({
+      const workouts = await prisma.workouts.findMany({
          include: {
             workout_applied_tags: {
-               include: {
-                  workout_tags: true
+               select: {
+                  workout_id: true,
+                  tag_id: true
                }
             },
             exercises: {
@@ -71,18 +76,14 @@ export async function fetchWorkouts(userId: string): Promise<Workout[]> {
             }
          },
          where: {
-            user_id: userId
+            user_id: user_id
          },
          orderBy: {
             date: "desc"
          }
       });
 
-      const formattedWorkouts = workoutsWithTags.map((workout) => {
-         return formatWorkout(workout);
-      });
-
-      return formattedWorkouts;
+      return workouts.map((workout) => formatWorkout(workout));
    } catch (error) {
       return [];
    }
@@ -92,11 +93,10 @@ export async function addWorkout(
    workout: Workout,
 ): Promise<VitalityResponse<Workout>> {
    try {
-      // Validate the feedback form first
-      const fields = workoutsSchema.safeParse(workout);
+      const fields = newWorkoutSchema.safeParse(workout);
 
       if (!fields.success) {
-         return sendErrorMessage("Invalid workout tag fields",
+         return sendErrorMessage("Invalid workout fields",
             fields.error.flatten().fieldErrors,
          );
       }
@@ -105,22 +105,19 @@ export async function addWorkout(
       const newWorkout = await prisma.workouts.create({
          data: {
             user_id: workout.user_id,
-            title: workout.title,
+            title: workout.title.trim(),
             date: workout.date,
-            image: workout.image,
-            description: workout.description,
+            description: workout.description?.trim(),
+            image: workout.image?.trim(),
             workout_applied_tags: {
-               create: workout.tagIds.map((tagId: string) => {
-                  return {
-                     tag_id: tagId
-                  };
-               })
+               create: workout.tagIds.map((tagId: string) => ({ tag_id: tagId }))
             }
          },
          include: {
             workout_applied_tags: {
-               include: {
-                  workout_tags: true
+               select: {
+                  workout_id: true,
+                  tag_id: true
                }
             },
             exercises: {
@@ -152,14 +149,16 @@ export async function updateWorkout(
          );
       } else {
          // Fetch existing tags first for data integrity
-         const existingWorkout = await prisma.workouts.findUnique({
+         const existingWorkout = await prisma.workouts.findFirst({
             where: {
-               id: workout.id
+               id: workout.id,
+               user_id: workout.user_id
             },
             include: {
                workout_applied_tags: {
-                  include: {
-                     workout_tags: true
+                  select: {
+                     workout_id: true,
+                     tag_id: true
                   }
                },
                exercises: {
@@ -173,44 +172,43 @@ export async function updateWorkout(
             }
          });
 
-         // Extract existing tag IDs
-         const existingAppliedTags: Set<string> = new Set(existingWorkout?.workout_applied_tags.map((tag) => tag.tag_id) || []);
+         if (!existingWorkout) {
+            return sendErrorMessage(
+               "Workout does not exist based on user and/or workout ID",
+               null
+            );
+         }
 
-         // Determine tags to connect and disconnect
-         const newAppliedTags: Set<string> = new Set(workout.tagIds);
-
-         const tagsToRemove: string[] = Array.from(existingAppliedTags).filter(
-            (id) => !newAppliedTags.has(id),
-         );
-         const tagsToAdd: string[] = Array.from(newAppliedTags).filter(
-            (id) => !existingAppliedTags.has(id),
-         );
+         const { adding, removing } = await getAppliedWorkoutTagUpdates(existingWorkout, workout);
 
          const updatedWorkout = await prisma.workouts.update({
-            where: { id: workout.id },
+            where: {
+               id: workout.id,
+               user_id: workout.user_id
+            },
             data: {
-               title: workout.title,
-               description: workout.description,
+               title: workout.title.trim(),
                date: workout.date,
-               image: workout.image,
+               description: workout.description?.trim(),
+               image: workout.image?.trim(),
                workout_applied_tags: {
                   // Disconnect existing applied tags
                   deleteMany: {
-                     tag_id: { in: tagsToRemove }
+                     tag_id: { in: removing }
                   },
                   // Create new applied tags
                   createMany: {
-                     data: tagsToAdd.map((tagId: string) => ({
+                     data: adding.map((tagId: string) => ({
                         tag_id: tagId
                      }))
                   }
                }
             },
-
             include: {
                workout_applied_tags: {
-                  include: {
-                     workout_tags: true
+                  select: {
+                     workout_id: true,
+                     tag_id: true
                   }
                },
                exercises: {
@@ -234,17 +232,38 @@ export async function updateWorkout(
    }
 }
 
-export async function removeWorkouts(
+export async function deleteWorkouts(
    workouts: Workout[],
+   user_id: string,
 ): Promise<VitalityResponse<number>> {
+   const errors = {};
+
+   if (!uuidSchema("user", "required").safeParse(user_id).success) {
+      errors["user_id"] = ["ID for user must be in UUID format"];
+   }
+
    try {
-      const ids: string[] = workouts.map((workout: Workout) => workout.id);
+      const ids: string[] = [];
+
+      for (const workout of workouts) {
+         if (!uuidSchema("workout", "required").safeParse(workout.id).success) {
+            errors["id"] = ["ID for all workouts must be in UUID format"];
+            break;
+         }
+
+         ids.push(workout.id);
+      }
+
+      if (Object.keys(errors).length > 0) {
+         return sendErrorMessage("Invalid workout ID fields", errors);
+      }
 
       const response = await prisma.workouts.deleteMany({
          where: {
             id: {
                in: ids
-            }
+            },
+            user_id: user_id
          }
       });
